api/requirements.txt
fastapi[standard]>=0.110.0
langchain-core>=0.1.30
langgraph>=0.0.26
groq>=0.4.2
pydantic>=2.6.0
python-dotenv>=1.0.1
uvicorn>=0.28.0